Fix command line to simplify bootstrap

+++ mnb-main/mnb/climain.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import argparse
+from inspect import signature
 from pathlib import PureWindowsPath
 
 import docker
@@ -39,25 +40,33 @@ def showplan(ns, plan_builder):
     p = mkplan(ns, plan_builder)
     p.show()
 
-def scripts(ns, _plan_builder):
-    if not ns.rootabspath:
+def scripts(ns):
+    if ns.dev_mode:
         src_dir_path = Path("scripts")
         dst_dir_path = Path(".")
     else:
         src_dir_path = PosixPath("/mnb/lib/scripts")
         dst_dir_path = PosixPath("/mnb/run")
-    for name in ["mnb", "mnb.cmd"]:
+    for name, overwrite, setx in [("mnb", True, True),
+                                  ("mnb.cmd", True, False),
+                                  ("mnb-plan.py", False, False)]:
+        if (dst_dir_path / name).exists() and not overwrite:
+            continue
         with (src_dir_path / name).open("r") as i:
             if dst_dir_path.is_dir():
                 with (dst_dir_path / name).open("w") as o:
                     c = i.read().replace("MNB_VERSION", MNB_VERSION)
                     o.write(c)
+                if setx:
+                    Path(dst_dir_path / name).chmod(755)
 
-def main(plan_builder):
+def main():
     parser = argparse.ArgumentParser(prog='mnb')
     parser.add_argument('--rootabspath', nargs='?', help="Absolute path to working context on host machine")
     parser.add_argument('--windows-host', action='store_true', help="host machine is running Windows (by default Unix-like system is assumed)")
     parser.add_argument('--argparse-test', action='store_true', help="dry run, print command-line parsing result")
+    parser.add_argument('--plan-file', default="mnb-plan.py", help="path to plan description")
+    parser.add_argument('--dev-mode', action='store_true', help="Development mode (run outside of a container)")
     parser.set_defaults(func=update)
     subparsers = parser.add_subparsers()
 
@@ -78,4 +87,20 @@ def main(plan_builder):
         print(ns)
         exit(0)
     else:
-        ns.func(ns, plan_builder)
+        n_args = len(signature(ns.func).parameters)
+        if n_args == 1:
+            ns.func(ns)
+        elif n_args == 2:
+            plan_file_path = Path(ns.plan_file)
+            if not plan_file_path.exists():
+                raise Exception("Plan file %s does not exist" % plan_file_path)
+            plan_env = dict()
+            exec(plan_file_path.open("r").read(), plan_env)
+            BUILD_PLAN = 'build_plan'
+            if not BUILD_PLAN in plan_env:
+                raise Exception("Function %s not defined in plan file %s" % (BUILD_PLAN, plan_file_path))
+            plan_builder = plan_env[BUILD_PLAN]
+            ns.func(ns, plan_builder)
+
+if __name__=="__main__":
+    main()
